+++ controller/researchResultsController.js
@@ -124,224 +124,6 @@ async function researchResultDelete(req, res, next) {
     }
 }
 
-// Patent
-async function researchResultsPatent(req, res, next) {
-    let token = req.session.user;
-    var queryPage = req.query.page;
-    var querySearch = req.query.schKeyword;
-    var parameters = {
-        "type": 'patent',
-        "page": queryPage,
-        "search": querySearch,
-        "name": 'vistors'
-    }
-    try {
-        console.log(parameters);
-        const permission = await jwtmiddle.jwtCerti(token);
-        const count_data = await counterDAO.findCount(parameters);
-        const db_data = await researchResultsDAO.researchResults_selectPatent(parameters)
-        res.render('research_results/researchResultsMain', { db_data, permission, parameters, dayjs, count_data });
-    } catch (error) {
-        res.send("<script>alert('" + error + "');location.href='/';</script>")
-    }
-}
-async function researchResultsPatentDetail(req, res, next) {
-    let token = req.session.user;
-    var queryNum = req.query.num;
-    var parameters = {
-        "rrid": queryNum,
-        "name": 'vistors'
-    };
-    try {
-        const permission = await jwtmiddle.jwtCerti(token);
-        const count_data = await counterDAO.findCount(parameters);
-        const db_data = await researchResultsDAO.researchResults_selectDetailPatent(parameters);
-        res.render('research_results/researchResultsDetailPatent', { db_data, permission, parameters, dayjs, count_data });
-    } catch (error) {
-        res.send("<script>alert('" + error + "');location.href='/';</script>")
-    }
-}
-async function researchResultsPatentUpdate(req, res, next) {
-    try {
-        let token = req.session.user
-        let queryNum = req.query.num;
-        let parameters = {
-            "rrid": queryNum,
-            "name": 'vistors'
-        }
-        if (token == undefined) throw "Parameter ERR."
-        const permission = await jwtmiddle.jwtCerti(token)
-        if (permission.userRole >= 5) throw "권한이없습니다."
-        const count_data = await counterDAO.findCount(parameters);
-        const db_data = await researchResultsDAO.researchResults_selectDetailPatent(parameters);
-        console.log(db_data)
-        res.render('research_results/researchResultsModifyPantent', { db_data, count_data, permission })
-    } catch (error) {
-        res.send("<script>alert('" + error + "');location.href='/';</script>")
-    }
-}
-async function researchResultsPatentDelete(req, res, next){
-    try {
-        let token = req.session.user
-        let queryNum = req.query.num;
-        let parameters = {
-            "rrid": queryNum,
-        }
-        if (token == undefined) throw "Parameter ERR."
-        const permission = await jwtmiddle.jwtCerti(token)
-        if (permission.userRole >= 5) throw "권한이없습니다."
-        const db_data = await researchResultsDAO.researchResults_DeletePatent(parameters);
-        res.redirect('/research/results/patent?schKeyword=&page=1');
-    } catch (error) {
-        res.send("<script>alert('" + error + "');location.href='/';</script>")
-    }
-}
-
-
-// Treatise
-async function researchResultsTreatise(req, res, next) {
-    let token = req.session.user;
-    var queryPage = req.query.page;
-    var querySearch = req.query.schKeyword;
-    var parameters = {
-        "type": 'treatise',
-        "page": queryPage,
-        "search": querySearch,
-        "name": 'vistors'
-    }
-    try {
-        const permission = await jwtmiddle.jwtCerti(token);
-        const count_data = await counterDAO.findCount(parameters);
-        const db_data = await researchResultsDAO.researchResults_selectTreatise(parameters)
-        res.render('research_results/researchResultsMain', { db_data, permission, parameters, dayjs, count_data });
-    } catch (error) {
-        res.send("<script>alert('" + error + "');location.href='/';</script>")
-    }
-}
-async function researchResultsTreatiseDetail(req, res, next) {
-    let token = req.session.user;
-    var queryNum = req.query.num;
-    var parameters = {
-        "rrid": queryNum,
-        "name": 'vistors'
-    };
-    try {
-        const permission = await jwtmiddle.jwtCerti(token);
-        const count_data = await counterDAO.findCount(parameters);
-        const db_data = await researchResultsDAO.researchResults_selectDetailTreatise(parameters);
-        res.render('research_results/researchResultsDetailTreatise', { db_data, permission, parameters, dayjs, count_data });
-    } catch (error) {
-        res.send("<script>alert('" + error + "');location.href='/';</script>")
-    }
-}
-async function researchResultsTreatiseUpdate(req, res, next) {
-    try {
-        let token = req.session.user
-        let queryNum = req.query.num;
-        let parameters = {
-            "rrid": queryNum,
-            "name": 'vistors'
-        }
-        if (token == undefined) throw "Parameter ERR."
-        const permission = await jwtmiddle.jwtCerti(token)
-        if (permission.userRole >= 5) throw "권한이없습니다."
-        const count_data = await counterDAO.findCount(parameters);
-        const db_data = await researchResultsDAO.researchResults_selectDetailTreatise(parameters);
-        console.log(db_data)
-        res.render('research_results/researchResultsModifyTreatise', { db_data, count_data, permission })
-    } catch (error) {
-        res.send("<script>alert('" + error + "');location.href='/';</script>")
-    }
-}
-async function researchResultsTreatiseDelete(req, res, next){
-    try {
-        let token = req.session.user
-        let queryNum = req.query.num;
-        let parameters = {
-            "rrid": queryNum,
-        }
-        if (token == undefined) throw "Parameter ERR."
-        const permission = await jwtmiddle.jwtCerti(token)
-        if (permission.userRole >= 5) throw "권한이없습니다."
-        const db_data = await researchResultsDAO.researchResults_DeleteTreatise(parameters);
-        res.redirect('/research/results/treatise?schKeyword=&page=1');
-    } catch (error) {
-        res.send("<script>alert('" + error + "');location.href='/';</script>")
-    }
-}
-
-// Announcement
-async function researchResultsAnnouncement(req, res, next) {
-    let token = req.session.user;
-    var queryPage = req.query.page;
-    var querySearch = req.query.schKeyword;
-    var parameters = {
-        "type": 'announcement',
-        "page": queryPage,
-        "search": querySearch,
-        "name": 'vistors'
-    }
-    try {
-        const permission = await jwtmiddle.jwtCerti(token);
-        const count_data = await counterDAO.findCount(parameters);
-        const db_data = await researchResultsDAO.researchResults_selectAnnouncement(parameters)
-        res.render('research_results/researchResultsMain', { db_data, permission, parameters, dayjs, count_data });
-    } catch (error) {
-        res.send("<script>alert('" + error + "');location.href='/';</script>")
-    }
-}
-async function researchResultsAnnouncementDetail(req, res, next) {
-    let token = req.session.user;
-    var queryNum = req.query.num;
-    var parameters = {
-        "rrid": queryNum,
-        "name": 'vistors'
-    };
-    try {
-        const permission = await jwtmiddle.jwtCerti(token);
-        const count_data = await counterDAO.findCount(parameters);
-        const db_data = await researchResultsDAO.researchResults_selectDetailAnnouncement(parameters);
-        res.render('research_results/researchResultsDetailAnnouncement', { db_data, permission, parameters, dayjs, count_data });
-    } catch (error) {
-        res.send("<script>alert('" + error + "');location.href='/';</script>")
-    }
-}
-async function researchResultsAnnouncementUpdate(req, res, next) {
-    try {
-        let token = req.session.user
-        let queryNum = req.query.num;
-        let parameters = {
-            "rrid": queryNum,
-            "name": 'vistors'
-        }
-        if (token == undefined) throw "Parameter ERR."
-        const permission = await jwtmiddle.jwtCerti(token)
-        if (permission.userRole >= 5) throw "권한이없습니다."
-        const count_data = await counterDAO.findCount(parameters);
-        const db_data = await researchResultsDAO.researchResults_selectDetailAnnouncement(parameters);
-        console.log(db_data)
-        res.render('research_results/researchResultsModifyAnnouncement', { db_data, count_data, permission })
-    } catch (error) {
-        res.send("<script>alert('" + error + "');location.href='/';</script>")
-    }
-}
-async function researchResultsAnnouncementDelete(req, res, next){
-    try {
-        let token = req.session.user
-        let queryNum = req.query.num;
-        let parameters = {
-            "rrid": queryNum,
-        }
-        if (token == undefined) throw "Parameter ERR."
-        const permission = await jwtmiddle.jwtCerti(token)
-        if (permission.userRole >= 5) throw "권한이없습니다."
-        const db_data = await researchResultsDAO.researchResults_DeleteAnnouncement(parameters);
-        res.redirect('/research/results/announcement?schKeyword=&page=1');
-    } catch (error) {
-        res.send("<script>alert('" + error + "');location.href='/';</script>")
-    }
-}
-
 module.exports = {
     researchResults,
     researchResultsDetail,
@@ -349,20 +131,5 @@ module.exports = {
     researcResultWrite,
     researcResultWriteP,
     // researchResultDelete,
-
-    researchResultsPatent,
-    researchResultsPatentDetail,
-    researchResultsPatentUpdate,
-    researchResultsPatentDelete,
-
-    researchResultsTreatise,
-    researchResultsTreatiseDetail,
-    researchResultsTreatiseUpdate,
-    researchResultsTreatiseDelete,
-
-    researchResultsAnnouncement,
-    researchResultsAnnouncementDetail,
-    researchResultsAnnouncementUpdate,
-    researchResultsAnnouncementDelete,
 
 }
