+++ views/research_results/researchResultsMain.ejs
@@ -94,7 +94,7 @@
                             <%}%>
                         </div>
                       </div>
-                      <div style="text-align: right; padding-right:6%"><a href="/research/results/write" type="button"
+                      <div style="text-align: right; padding-right:6%"><a href="/research/results/<%=parameters.type%>/write" type="button"
                         class="btn btn-outline-secondary">작성하기</a>
                 </div>
                 </div>
