+++ views/research_results/researchResultsWritePatent.ejs
@@ -2,7 +2,7 @@
     <link rel="stylesheet" href="/stylesheets/research_fields/researchFieldsMain.css" />
     <link rel="stylesheet" href="/stylesheets/research_fields/researchFieldsDetail.css" />
     <section>
-        <form method="post" action="/research/results/write">
+        <form method="post" action="/research/results/patent/write">
             <div id="indexbg" class="mb-5"></div>
             <div id="indexbg_text">
                 <h1 class="fs-1">연구 분야</h1>
@@ -36,7 +36,7 @@
                                     <tr>
                                         <th>출원번호</th>
                                         <td>
-                                            <input style="width: 100%; border: none;" type="text">
+                                            <input style="width: 100%; border: none;" type="text" name="application_num">
                                         </td>
                                         <th class="align-middle">출원일자</th>
                                         <td>
