+++ views/research_results/researchResultsWriteTreatise.ejs
@@ -2,7 +2,7 @@
     <link rel="stylesheet" href="/stylesheets/research_fields/researchFieldsMain.css" />
     <link rel="stylesheet" href="/stylesheets/research_fields/researchFieldsDetail.css" />
     <section>
-        <form method="post" action="/research/results/write">
+        <form method="post" action="/research/results/treatise/write">
             <div id="indexbg" class="mb-5"></div>
             <div id="indexbg_text">
                 <h1 class="fs-1">연구 분야</h1>
